implemented jwt pr changes
