+++ .gitignore
@@ -2,3 +2,4 @@ venv/
 .idea/
 .flaskenv
 __pycache__/
+app/__pycache__/
