+++ requirements.txt
@@ -55,10 +55,6 @@ PyJWT==2.3.0
 pymacaroons==0.13.0
 PyMySQL==1.0.2
 PyNaCl==1.5.0
-pyparsing==2.4.7
-pyRFC3339==1.1
-python-dateutil==2.8.1
-pytz==2022.1
 pyxdg==0.27
 PyYAML==5.4.1
 reportlab==3.6.8
